requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
crispy-bootstrap4==2022.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-admin-rangefilter==0.10.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
flake8==6.0.0
flake8-builtins==2.1.0
flake8-import-order==0.18.2
flake8-print==5.0.0
gprof2dot==2022.7.29
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
kombu==5.3.1
MarkupPy==1.14
marshmallow==3.20.1
matplotlib-inline==0.1.6
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.3.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
